Move non-lead tests to the same file as other differentiability tests
